requirements.txt
opencv-contrib-python==3.4.5.20
opencv-python==3.4.5.20
pathlib==1.0.1
PyYAML==3.13
requests==2.21.0
numpy==1.15.4